Changes to be committed: 	renamed:    operations/notify.py -> transport/notify.py 	modified:   transport/session.py
